[Changed] implementation as noted in review.
- [Added] travis CI config file

+++ src/rust/include/rregex.h
@@ -3,5 +3,5 @@
 struct Regex;
 
 extern struct Regex const * regex_new(char * const regex);
-extern uint32_t regex_matches(struct Regex const * const exp, uint8_t * const str, uint64_t str_len);
+extern uint64_t regex_matches(struct Regex const * const exp, uint8_t * const str, uint64_t str_len);
 extern void regex_free(struct Regex const * const exp);
